Phase 17: security hardening, cargo-deny, and CI/docs closeout
- Bump wasmtime 24 -> 43 to clear six open RustSec advisories (several
  Wasm-sandbox-escape bugs) surfaced by the new cargo-deny check.
- Fix RenderGraph::execute: nodes were never granted capability access
  to the resources they create, so every render_frame call failed
  security validation, including the triangle_and_path milestone
  example. CapabilityGuard's grant sets now live behind RefCell so a
  shared &SecurityContext can grant during graph execution.
- Allow wildcard path deps for internal workspace crates in deny.toml.
- Fix FfiBridge table-element-limit type mismatch (u32 -> usize).

+++ crates/tpt-chora-compat/src/ffi_bridge.rs
@@ -13,7 +13,7 @@ const FUEL_PER_CALL: u64 = 50_000_000;
 /// filesystem/network; this bounds how much host memory it can claim via
 /// `memory.grow`, so it can't OOM the host process either.
 const MAX_MEMORY_BYTES: usize = 256 * 1024 * 1024;
-const MAX_TABLE_ELEMENTS: u32 = 10_000;
+const MAX_TABLE_ELEMENTS: usize = 10_000;
 
 pub struct FfiBridge {
     engine: Engine,
